lxd/request: Never get existing `Info` when setting up request context.
Also rename to `InitContextInfo`.

When setting up the context info, we should never use the existing value
verbatim. This is because the UI multiplexes requests over HTTP2, and
per https://pkg.go.dev/net/http#Request.Context, this context is
cancelled when the clients connection is cancelled. This means that
multiple requests are made on the same connection and they can all
share a context.

Happily, when we set the Info in the request context, we perform a
shallow clone of the Request and pass it down to further handlers. This
means that we can set an empty value when the request is received and
rely on it being correct when we later retrieve it because it is set
higher in the context tree (this behaviour is also relied upon by mux
for HTTP vars).

Signed-off-by: Mark Laing <mark.laing@canonical.com>