api: Support booting without EFI / OVMF
VMs sometimes don't need the full feature-set of OVMF. Provide an
option to select an alternative (based on U-Boot) which can boot more
quickly and with a smaller attack surface.

The target audience is people who want  a lighter-weight VM that boots
almost as fast as a container, but is still a VM. Future additions may
include Firecracker, Coconut SVSM, etc.

Signed-off-by: Simon Glass <simon.glass@canonical.com>