lxd/auth/oidc: Allow different cluster members to begin and end UI auth flow.
This commit uses the new `RequestAwareCookieHandler` available in
zitadel/oidc to derive cookie encryption keys from the cluster private
key, as we do for other OIDC cookies.

To do this, a `login_id` cookie is set on calls to `/oidc/login`. When
the caller is redirected to `/oidc/callback`, the same `login_id`
cookie is found and used to derive the same encryption keys.

The `login_id` cookie is deleted after the callback completes.

Closes #13644

Signed-off-by: Mark Laing <mark.laing@canonical.com>