lxd/auth/oidc: Remove config expiry logic.
Config expiration logic was added to the Verifier so that it would
regularly rotate cookie encryption keys. However, cookie encryption keys
are now derived on a per-login basis, so there is no reason to
reconfigure the relying party (which means we no longer need to
re-perform discovery or fetch JWKs).

This commit removes most config expiry logic. However, we do still need
to reconfigure the verifier on-demand if any proxy configuration is
changed. A comment is added to this effect.

Signed-off-by: Mark Laing <mark.laing@canonical.com>