lxd/instance/drivers/driver/qemu: Update addDriveDirConfig to use a busAllocater that uses per-LXD device multi-function groups
This fixes an issue where custom directory share disk devices were using the same multi-function slot as the config drive.

This worked fine when the VM was started, but if the custom disk was then hot-unplugged then the entire slot was emptied,
which also removed the config drive devices.

This change causes custom directory share disk devices to use their own slot, like it does when they are hotplugged.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>