lxd/auth/oidc: Simplify Auth logic.
Changes the signature of `authenticateIDToken` to accept a
http.ResponseWriter and a http.Request. It can then get the ID and
refresh tokens itself. This simplifies logic in the `Auth` function,
which can now perform a simple check to decide if it should authenticate
a bearer token, or authenticate an ID/refresh token.

Signed-off-by: Mark Laing <mark.laing@canonical.com>