lxd/apparmor: avoid shadowing `supported` variable in `parserSupports()`
Aleks noticed that when I switched to using `:=` to deal with the then new
`err` variable, I also accidentally shadowed the `supported` variable. This meant
that on first invocation of `parserSupports()`, it would always return "not
supported" (`false`) and on subsequent calls it would return if the feature was
actually supported or not.

Closes https://github.com/canonical/lxd/issues/16114
Fixes commit aa7fbbcfc402ba48c6be8deed6939c6009de9769

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>