lxd/cluster/config: Remove ClusterTx from Config.
A pointer to `config.Config` is saved to the Daemon struct but
the transaction it holds is not valid outside of the initial
transaction where it was loaded. This is pretty strange. Instead,
let's pass a transaction only into the methods that require it.

Signed-off-by: Mark Laing <mark.laing@canonical.com>
(cherry picked from commit dc8272432ca8dd6b9b8e144647dde71f862afca2)