lxd/storage/backend: Add custom volume discovery when checking instance volumes
Iterate over all the custom volumes attached to an instance and identify the ones which don't yet exist in the DB.

To follow the protocol we pack each custom volume in its own backup config containing a single volume.
This repeats the behavior already performed by detectUnknownCustomVolume.

In case the instance also needs to be restored we delete all the custom volumes from the list so we don't
accidentally report the custom volumes twice to the caller.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>