lxd/api_internal_recover: Unpack the unknown volumes returned from ListUnkownVolumes
When listing unknown volumes from a pool, we have to make sure that they are all placed into their respective pool's buckets.
As we now support the discovery of custom volumes from the list of volumes attached to instances, a call to
ListUnknownVolumes might not only yield custom volumes on this exact pool but also from other pools which actually host the custom volumes.

In addition we try to enhance the information from already discovered volumes with the enriched information we get from an instances backup config.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>