lxd/storage/connectors: fix NVMe Discovery() to filter out not valid NQNs
We should be more picky about NQNs we take from .Discovery() method
as in case of NVMe it can return not only "nvme subsystem" NQN types,
but also "discovery subsystem referral" and "current discovery subsystem".
These two are non valid for our use case, as we need an NQN we can use
to connect to storage array.

Signed-off-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>