lxd/storage/backend: Append instance vol only after validation
This technically has no effect because if the snapshot validation fails, the function returns an
error which causes the just inserted backup config (to the global map) to become irrelevant.

But it supports the readability of the func.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>
(cherry picked from commit 7d9819aba3b45e628e1402c551a30dba2ff1037e)