lxd/storage/drivers/pure: Ensure volume is unmapped after changing quota
Compared to PowerFlex, the Pure driver is taking a different approach for unmap after volume size changes.
SetVolumeQuota can be invoked in two cases, when the volume is already mapped and when the volume is not mapped.
In the latter case the volume should be unmapped again which won't happen if the cleanup hook from
getMappedDevPath is only added to the reverter which is cleared right before exiting the function.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>
(cherry picked from commit 07b1821ad34d12f5696fae031add6c956201b074)
Signed-off-by: Din Music <din.music@canonical.com>