lxd/storage/drivers/powerflex: Replace powerflex.clone_copy with powerflex.snapshot_copy
This ensures the 'clone_copy' setting isn't confused with the one present for the ZFS and Ceph RBD drivers.
See https://github.com/canonical/lxd/issues/15923.

Instead a new setting 'snapshot_copy' is introduced which defaults to 'false'.
This ensures the PowerFlex driver by default creates a non-sparse volume copy but may use the snapshot functionality
when creating a new volume in case 'snapshot_copy' is set to true.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>