doc: Add DevLXD API specification.
The specification is currently handwritten. It was not possible to
generate a separate specification for DevLXD without modifying the
existing main API spec.

Please advise as to where this should be placed. If I placed it
anywhere other than `doc/.sphinx/_static` it was not included in the
doc build and I couldn't reference it.

Signed-off-by: Mark Laing <mark.laing@canonical.com>