lxd/main_init_interactive: Display cluster member config key defaults
This helps for most config keys which usually are the same on all cluster members.
Only for local storage pools, the 'source' config key has to be changed when adding fresh cluster members.
When recovering cluster members with existing local storage pools, the default for 'source' can be used which
should make the recovery process more supportive.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>