lxd/storage/backend: Allow reuse of pools only existing in storage
When trying to create a pool, it might be that LXD isn't yet aware of it but it already exists on the underlying storage.
This is the case when recovering lost cluster members which still have the underyling storage present but lost LXD.
Also this might be the case when swapping out physical disks to another machine which is then used to replace the original cluster member.

Right before trying to create the pool on the underlying storage, we try if we can successfully mount it.
If that is the case we can be sure that the pool already exists and we just have to re-create its DB
entries in LXDs database.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>