github: use actions/setup-python only for non-PRs
This action was introduced to save `pip`'s cache to speed up the building of
the docs. However, since commit a7828ec63c0dddf207c11bc029508d8bc3a09be1, when
running `make dist` in the the `Code`, we skip building the docs if we running
as part of a PR. This was made as a speed optimization too.

Since we have no use for Python during `make dist` on PR test runs, might as
well not even invoke the action.

This avoids https://github.com/actions/setup-python/issues/1169

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>