lxd: Further restrict setting of PKI protocol.
Previously, we relied on the authorizer to check if the identity was in
the cache during each access check. Now that the identity is set in the
request context, we need to restrict this further so that PKI is only
set when `core.trust_ca_certificates` is true, the client certificate is
signed by the CA, and the certificate is not present in the identity
cache.

Signed-off-by: Mark Laing <mark.laing@canonical.com>