lxd/storage/drivers: Add driver level patches
These allow creating a bridge between the drivers package and the patch system.
To not require exporting driver specific functions (e.g. HTTP clients), a custom patch function can be created
on the driver level which allows calling those internal functions.

On the other hand the caller can query the database to collect the necessary information which is not possible
from the perspective of the storage driver.

The PatchUpdatePowerFlexSnapshotPrefix function expects a list of snapshots which require patching of their actual
name to include a dedicated snapshot prefix which is not applicable for all volumes of type snapshot.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>
(cherry picked from commit 52716c9e2679c299d1c29bacabc2913cb7f62251)