lxd: Further restrict PKI protocol.
We rely on the authorizer to check if the identity is in the cache during
each access check. We will soon stop doing this and instead add the
identity to the request context. So we need to restrict this further so
that PKI is only set when `core.trust_ca_certificates` is true, the client
certificate is signed by the CA, and the certificate is not present in the
identity cache.

Signed-off-by: Mark Laing <mark.laing@canonical.com>
(cherry picked from commit 3315aed10008c51492e44cdb63f037e162a917a9)