lxd: Improve error handling when reporting entitlements.
This is a follow up of https://github.com/canonical/lxd/pull/16223#discussion_r2287314061.

When the identity type is not present, it is not necessarily an
internal error. It is a bad request if the calling identity is
not fine-grained. This checks if the identity is a global admin
(Unix, unrestricted TLS, CA signed cert with
`core.trust_ca_certificates=true`) before getting the identity
type and checking if it is fine-grained.

Signed-off-by: Mark Laing <mark.laing@canonical.com>