test/suites/clustering: Add test_clustering_recovery
Test a scenario in which there is a local storage pool across three cluster members.
In case one cluster member dies, we want to be able to recover an already existing storage pool on any member
when adding it back to the cluster.

In this scenario one cluster member uses a loop device for its ZFS pool which will survive the recreation of the cluster member.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>