lxd/request: Handle devlxd protocol in setIdentity.
When setting the Requestor, we previously expected DevLXD requests to
always be untrusted. Now that guests can authenticate with a token, we
need to get these identities from the cache. To do so, we use the
"bearer" authentication method instead of the "devlxd" protocol. This
is the same as we currently do for the "cluster" protocol (which is
shorthand for "tls" with identity type "Server certificate"). This
assumption could be problematic if any other authentication method
is ever allowed over DevLXD.

Signed-off-by: Mark Laing <mark.laing@canonical.com>