lxd/cluster/gateway: Use outer context for TLS dial
When using only the outer context's deadline as timeout for the actual TLS dial,
we ignore cases in which the outer context gets cancelled which should also cancel the TLS dial.

In case the target addr is offline, the dial function will wait until the timeout (outer context's deadline)
instead of returing early in case the context already got cancelled.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>