lxd/auth/drivers: Enrich logging for OpenFGA errors.
Previously, if an error occurred during an OpenFGA Check or ListObjects
request, we attempted to get the underlying error and returned that to
the end user.

This is problematic because we don't know what the contents of that
error will be. For example, it may contain information about tuples that
the user is not related to (can't access).

This commit changes error handling such that:
1. Only the public OpenFGA error is returned to the end user (this will
   likely just be "internal server error").
2. The underlying error is extracted an logged.
3. The callsite is logged.
4. All other details about the request that failed are logged, including
   the identity id, protocol, entitlements, and requested URL.

The callsite is logged because these errors are always unexpected, so it
helps to understand the request flow when fixing issues.

Signed-off-by: Mark Laing <mark.laing@canonical.com>
(cherry picked from commit 1eeaf84692008bce74f11af2416b41043b4dd23a)