lxd/auth/drivers: Remove 5s timeout on fine-grained auth checks.
This timeout was added with the intention of surfacing access checks
that are taking too long to perform. However, we can't help access
checks taking a long time when the cluster is under heavy load, as we
need to wait for a lock on the database. This commit removes the timeout
and instead relies on the global 10s transaction timeout.

Signed-off-by: Mark Laing <mark.laing@canonical.com>