lxd/events: Swap the permission checker for a more general filter function.
The filter argument is now a `func(logger.Logger, api.Event)`. The
logger is passed in so that the filter can log any errors that occur.
The filter function must not use the global logger, otherwise this sends
an event to the event server while the current event is being processed,
which causes a deadlock.

Signed-off-by: Mark Laing <mark.laing@canonical.com>