lxd: Handle view access for operations with no project.
Previously we were checking if the caller had "can_view_operations"
against a project with an invalid URL (e.g. no project name, so
"/1.0/projects/"). This commit makes handling view access on these
operations explicit.

Since these are server level operations, they are considered to be for
administrators only. Therefore I've restricted view access to only those
with `can_view_operations` on `server`. This is the same entitlement
that is used to filter out server level operations in the events websocket.

Signed-off-by: Mark Laing <mark.laing@canonical.com>