lxd: Add access handling on GET /1.0/operations/{id}.
Previously, we relied on the fact that operation IDs are UUIDs and are
therefore not guessable. So if the user cannot list operations via
/1.0/operations, they can't get the operation. This does not account for
a bad actor acquiring an operation ID by other means.

Operations can contain metadata with e.g. URLs of entities that a user
is not supposed to be able to view. So it isn't sufficient to perform
access control based on what a bad actor is able to guess.

This commit updates access control in GET /1.0/operations/{id} to
perform the same check as when listing operations.

Signed-off-by: Mark Laing <mark.laing@canonical.com>
(cherry picked from commit eb4bf028b47fb1d0e8999425dac6b8a93bb4ffd2)