lxd: Remove default project from server level lifecycle events.
Events relating to the following entity types:
- `server` (configuration update)
- `cluster_group`
- `auth_group`
- `certificate`
- `identity`
- `identity_provider_group`
- `storage_pool`

Are not specific to a project, but lifecycle events are being
emitted in the default project. This means that e.g. a restricted
TLS client with access to the default project would be able to see
lifecycle events about identity creation via the events websocket.

With improvements to event filtering in #15991, we can make these
events require `can_view_events` on `server` by not setting a
project name.

Signed-off-by: Mark Laing <mark.laing@canonical.com>