lxd/instance/drivers/lxc: ignore clean stop of SFTP server
If a SFTP server was spawned and is still around when a container is stopped,
it will be `SIGTERM`'ed to do an orderly shutdown. This is perfectly normal so
it shouldn't cause any log message.

Fixes #15925

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>
(cherry picked from commit b301c05e2ba8316d228902683aea9c19d265d1d6)