lxd/storage/drivers/pure: fix block device unmap logic for iSCSI
We found that our previous approach with iSCSI volumes (in multipath mode)
unmap doesn't work, because we have a race condition with systemd-udevd which
tries to access process a device changes and as a part of it also tries to
access a block device /dev/dm-X. But as long as by that time we have removed
a vLUN from the array, this access attempt traps the process to get into
D-state. A clean solution is to:
1. remove a block device first
2. remove vLUN

Also, for multipath device we can't just remove the device, we have to use
a command "multipath -f /dev/dm-X" to do that. Because otherwise we also
at risk of stepping on our tail. Because to remove a multipath device we
need to remove it slaves (backing block devices), and once all of them
are removed then /dev/dm-X device becomes inaccessible too. And again,
we can race with udevd who can (and he is!) try to access it and get stuck.

Port of 6b6c112 ("lxd/storage/drivers/alletra: fix block device unmap logic for iSCSI").
Signed-off-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>
(cherry picked from commit a0419acae597d52faaa2d70a858ed3f5ff2774f1)