test/suites: Ensure empty string tests fail if command fails.
We set -e in the test suite to ensure an exit on error. Unfortunately,
as far as I can tell -e doesn't apply inside a test. The man page for
bash (https://man7.org/linux/man-pages/man1/bash.1.html) has:

> The shell does not exit if the
> command that fails is part of the command list
> immediately following a while or until reserved
> word, part of the test following the if or elif
> reserved words, part of any command executed in a
> && or || list except the command following the
> final && or ||, any command in a pipeline but the
> last (subject to the state of the pipefail shell
> option), or if the command's return value is being
> inverted with !.

Which to me indicates that it should be failing, but at least locally
for me it is not.

This means that any test that runs `[ "$(cmd)" = "" ]` is not erroring
out if cmd fails, and the test is asserting that the stdout from cmd
is empty. This is usually true if the command fails, as it will send
output to stderr instead.

To fix this, this commit adds `|| echo fail` to each case. If the
command fails, bash will execute OR statement causing the equality to
fail.

Signed-off-by: Mark Laing <mark.laing@canonical.com>