test/suites/basic: improve file transfer checks
The `lxc exec rm` cases were failing due to missing the leading `/` on the file
path to delete. Since `rm -f` was used, this was a silent failure.

It's not useful to `lxc exec` to `rm` or `cat` files so might as well test out
the `lxc file delete` and `lxc file pull` features instead.

Also cleanup the source file once done.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>