test/suites/container_devices_tpm: install `linux-modules-extra` to get the needed TPM module
This is installed opportunistically as this package is quite big:

* ~110MiB for the `-generic` (same as `-virtual`)

So it's best to pull it only if needed. This way, all the `cluster` test won't
need to.

The `-azure` kernel flavor is missing `CONFIG_TCG_VTPM_PROXY` so even the
`linux-modules-extra-azure` package doesn't ship the needed `.ko`.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>
(cherry picked from commit cfd2eedbc05715fcb66d007331d9a45abad08430)