test/suites: Update server config test for OIDC session expiry validation.
The test previously asserted that the core.auth_secret_expiry cannot be
set to one day. This is still true, but since the default OIDC session
expiry is one week, by default the auth secret expiry actually can't be
set to less than a week.

This commit updates the test to make assertions about the
oidc.session.expiry and interactions with core.auth_secret_expiry.

Signed-off-by: Mark Laing <mark.laing@canonical.com>