lxd/auth/drivers: Set list object max results to 10000.
Previously, calls to `(openfgav1.OpenFGAServiceServer).ListObjects`
limited the number of results to the default value of 1000 (see
https://github.com/openfga/openfga/blob/2a454e3772389af0e06f5472eb2d898c264cc732/pkg/server/config/config.go#L25).

This commit increases the limit to 10000 so as to always return all
data. It assumes that there are never more than 10000 of a LXD resource.

Note that increasing this limit might unnecessary increase memory usage
because OpenFGA creates a channel of this size when executing the query.

Closes #16614

Signed-off-by: Mark Laing <mark.laing@canonical.com>