test/main: add preflight check to `run_test`
This allows interleaving regular and snap-based tests like:

```
./main.sh server_info snap_lxd_user metrics snap_basic clustering_enable
```

and have the script prepare the right LXD daemon for each test.

Note: while supported, such interleaving is not optimal for speed due to the
setup time associated with the snap installation and removal. A better way
would be to group them so that all `snap_*` tests are run first or last:

```
./main.sh server_info metrics clustering_enable snap_lxd_user snap_basic
```

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>