daemon.start: improve Apparmor unprivileged userns/unconfined mediations handling
By persisting the changed settings into a `/run/sysctl.d` snippet we ensure
that any external tool reloading system rules will not undo our changes if and
only if LXD is running.

The rational for using a snippet in `/run` instead of `/etc` is to avoid
persisting the changes longer than needed if LXD removed/deactivated.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>