github: free up disk space for the `standalone` test group
This should reduce the likelyhood of running out of space with GH action
runners.

With the addition of the `devlxd_vm` test, large VMs are created onto pools
backed by `loop` files. So the test cleans up after itself but the backing file
is never shrunk after being inflated to accomodate the VM.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>