lxd/db/cluster: Return 404 if no project when getting config.
This function was added to optimise getting configuration for a single
project. This is because the generated "Get{X}Config" function load
configuration for all entities of the given type.

The generated "GetProjectConfig" function used to accept a project ID.
This meant the caller had to determine that the project existed before
getting the configuration.

The new function accepts a project name and performs a join. A
consequence of this change is that an empty configuration map is
returned if the project does not exist (it is not possible to
distinguish between a project with no configuration, or a project that
doesn't exist).

This is problematic when checking project feature flags. In many API
handlers, the feature flag check doubles as a project existence check.
In particular the images API is affected such that when the images
feature is checked for a project that does not exist, this function
returns an empty map, which is interpreted as `features.images=false`.
This causes the handler to think that the effective images project is
the default project. To the end user, this leads to:
```
lxc image list --project does-not-exist
```
being equal to
```
lxc image list --project default
```

This commit fixes the issue by removing the JOIN from the SQL query
and getting the project ID first - returning a 404 if it doesn't exist.

Signed-off-by: Mark Laing <mark.laing@canonical.com>