test/suites: Fix public image list assertions.
Changes to access handling for GET /1.0/images made it more consistent
with access handling for all other project level resources in LXD.

All callers can list images in the default project, if they are not
trusted or do not have access to the default project, they will only see
public images.

For other projects, a generic 404 is returned for untrusted callers and
fine-grained identities. A 403 is returned for restricted TLS clients
(consistent with other entity types).

Signed-off-by: Mark Laing <mark.laing@canonical.com>