lxcfs: make lxcfs.pidfd=true by default
No reason to not to use pidfd's these days. It is well tested
kernel API and it's better to switch to it at least on latest/edge
series.

This change makes no noticable functional changes for LXCFS, it just
enabled pidfd-based process tracking, instead of using stat() syscall
on the /proc/<pid> directory.

Signed-off-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>