doc/installing: manually run the `lxd` daemon with the `lxd` group
Reusing the `sudo` group is not standard and might have been convenient but
since the `lxd` group is created a few steps earlier, there is no point in
abusing the `sudo` group.

Also remove the `-E` argument for `sudo` as needed environment variables are
already passed to the `lxd` daemon already and `-E` is not implemented by
`sudo-rs`. Fixes https://github.com/canonical/lxd/issues/16962

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>