doc/howto/disaster_recovery_replication: Drop Pure until we were able to test with two clusters
Regular volume recovery works for Pure using 'lxd recover'.

Due to the availability of hardware, we haven't yet tested replication between two Pure clusters in order to replay the disaster recovery process.
Therefore we drop the link to the docs that might suggest it's supported until we have been able to perform tests.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>