lxd/storage/zfs: Rework zvol resolution logic
The previous logic was relying either on udev or on a slow crawl through
/dev using the zvol_id utility.

This was problematic not just because we could get into timeouts on
systems with a LOT of volumes, but also because it was inherently racy.
Udev works in the background, so entries may not show up immediately and
if something goes wrong during cleanup, we can be left with a symlink
pointing to the wrong /dev/zdXYZ device.

This is very similar to the LVM issue we fixed a little while back where
in the worst case scenario, it can lead to the wrong block device being
accessed, causing corruption or potential data leakage.

The new logic relies entirely on devtmpfs (not udev) and is optimized to
try and hit the correct /dev/zdXYZ device almost immediately, limiting
the performance impact on busy systems.

Closes https://github.com/lxc/incus/issues/1505

Signed-off-by: Stéphane Graber <stgraber@stgraber.org>
(cherry picked from commit 3a26837acde2f2b8d0612b55780feab5a28ee098)
Signed-off-by: Simon Deziel <simon.deziel@canonical.com>
License: Apache-2.0
(cherry picked from commit 9bb335f6842d68c5122b54149dfa9e399a90533a)