lxd/operations: Propagate context to operations.
This commit removes the `cancel.Canceller` from an `Operation` and changes
it back to a `context.Context` with a `context.CancelFunc`. This is so that
the `Operation` retains a context.

The context of the operation is the `Daemon` shutdown context plus key
information that may have been gained before the operation was started,
such as the `request.Requestor`.

Signed-off-by: Mark Laing <mark.laing@canonical.com>