lxd/db: Remove an instances volume when clearing a node
When clearing a node, it's important to not only remove the instance but also it's corresponding volume.
Otherwise we leave broken traces of the instance behind.

One issue this is causing is that you cannot anymore run 'lxd recover' to recover the resource (on remote storage)
because the corresponding volume DB entry is still around.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>
(cherry picked from commit db7bbb6762e1c0dd11416a4cd642c51873d16a2e)