lxd-agent: Check DevLXD connection manually
If DevLXD was previously disabled, when LXD Agent checks connection, the endpoint
is forbidden, because it was not updated in DB yet.
Therefore, to allow checking for security.devlxd flag in /1.0 endpoint, LXD agent
needs to consider Forbidden error as "successful connection". The flag will be
switched once LXD agent starts the DevLXD server.

Signed-off-by: Din Music <din.music@canonical.com>