lxd: Save connection in context for vsock API.
A connection can be found in the request context for all remote API
calls to LXD except for the VSock API. We can use this to perform some
assertions about when we expect a requestor to be present in the context
that is passed to an operation. If there is no connection, it is a
background operation.

Signed-off-by: Mark Laing <mark.laing@canonical.com>