client: Call mkdir instead of file stat.
When LXD is concurrently downloading images, it creates a local lock per
image. The lock is keyed on image fingerprint.

Unfortunately, we need to instantiate a client before creating said lock
because we need a client to resolve an image alias to a fingerprint to
key the lock on.

This makes the existing stat+mkdir racey because another routine could
create the directory after another routine has checked it's existence.

Closes #16687

Signed-off-by: Mark Laing <mark.laing@canonical.com>