doc: set html_baseurl dynamically
Replaces the statically set value for html_baseurl to use RTD
environment variable, or "/" if not available (for local builds). This
ensures that the HTML metadata canonical URL is set correctly, including
the version in the URL.

This change required also updating the sitemap_url_scheme to prevent
duplicating the version string, since the sitemap extension uses
html_baseurl to create the link.

References:
- https://docs.readthedocs.com/platform/stable/canonical-urls.html
- https://www.sphinx-doc.org/en/master/usage/configuration.html#confval-html_baseurl

Signed-off-by: Minae Lee <minae.lee@canonical.com>