lxd/network/acl: Don't get instance usage within project ACL loop.
The ACL UsedBy function gets a list of ACL names in the project, then
iterates over those ACLs to determine if the ACLs are "used by" each
other. This is because ACLs can reference other ACLs in the source and
destination fields.

Unfortunately, the logic that determined instance usage was also running
within this loop. This led to repeated instances in the used-by list,
with the number of repetitions matching the number of ACLs in the
project.

Closes #17011.

Signed-off-by: Mark Laing <mark.laing@canonical.com>
(cherry picked from commit d17ae59c859083f179b604357bcc60e746a7e066)