test/suites/clustering: Cleanup pool after cluster evacuation tests
This test leaves pools behind on the underlying storage as it does not use the 'dir' driver.
Only when using 'dir', the final call to the kill_lxd func will be able to cleanup the underlying storage.

We can fix this by manually triggering the cluster pool removal before cleanup.
This is also performed for all other tests that use a specific driver when forming the cluster.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>