lxd/operations: Validate using context rather than status.
The operation status is a consequence of the running and finished
statuses.

Additionally, this change allows an operation that is "cancelling" to be
updated.

This allows a run hook to, for example, change operation metadata after
the context is cancelled. This is useful for returning exit codes to the
end user. For example, if an exec websocket operation is deleted by an
administrator this will boot the user from the exec session with an
instance. By checking the context of the run hook, the exec websocket
can return a useful error to the user.

Signed-off-by: Mark Laing <mark.laing@canonical.com>