test/includes: Fix project argument for volume operation.
Noticed error logs while running the shutdown suite that read e.g.
'Failed creating operation: Failed finding ID of project "20s"'
which looked a bit odd.

I'm not certain why tests were not failing, but it might be a signal
that we need to look closer at the shutdown tests.

Signed-off-by: Mark Laing <mark.laing@canonical.com>