lxd/main_recover: Drop unknown pools concept
It's not anymore up to the user to define additional unknown pools.
Instead unknown pools are now indicated by the server as part of the recovery response.

In case other pools need to be recovered, a user will make use of the 'lxc storage create' command
which always works for both standalone LXD servers and clusters.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>