lxd/seccomp/seccomp: mknod: continue syscall for whiteouts
Whiteout devices are special and they are allowed to be created
by unprivileged user [1], cause they are just a placeholders useful
for overlayfs.

When mknod interception is enabled and we see that userspace wants
a whiteout, let's just do a syscall continue because native
kernel mknod() implementation is more reliable in some cases
we can't fix now.

For most cases, this change can be considered as a micro-optimization,
while for some others (like in #14849) it actually makes things to work.

[1] https://github.com/torvalds/linux/commit/a3c751a50fe6bbe50eb7622a14b18b361804ee0c

Fixes: #14849
Signed-off-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>