lxd/storage/drivers/cephfs: Already populate cephfs.path in FillConfig
This ensures when recovering a CephFS pool, 'cephfs.path' gets populated accordingly so a user can just specify 'source' as
it's done when creating a new pool.

Otherwise you either need to put 'source=foo/' for 'source=foo' and 'cephfs.path=foo'.
This is in line with the other drivers which populate all of the neccesary configuration as part of FillConfig.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>