test/suites/image: optimize `image_expiry`
When waiting for image expiry, instead of waiting for the command to fail and
then break, it now checks if it failed at the start of each iteration and break
immediately. This avoids waiting for ~up to 2 iterations if we just missed the
expiry.

Also, poll more often (every 0.5s) as we are just busy waiting after LXD scheduled cleanup.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>